Moved base & dummy routine classes to better packages
