+++ src/uk/ac/exeter/QCRoutines/routines/Dummy/DummyRoutine.java
@@ -1,12 +1,12 @@
-package uk.ac.exeter.QCRoutines.routines;
+package uk.ac.exeter.QCRoutines.routines.Dummy;
 
 import java.util.List;
 
-import uk.ac.exeter.QCRoutines.Routine;
-import uk.ac.exeter.QCRoutines.RoutineException;
 import uk.ac.exeter.QCRoutines.config.ColumnConfig;
 import uk.ac.exeter.QCRoutines.data.DataRecord;
 import uk.ac.exeter.QCRoutines.messages.Flag;
+import uk.ac.exeter.QCRoutines.routines.Routine;
+import uk.ac.exeter.QCRoutines.routines.RoutineException;
 
 public class DummyRoutine extends Routine {
