+++ src/uk/ac/exeter/QCRoutines/routines/FixedValue/FixedValueRoutine.java
@@ -3,13 +3,13 @@ package uk.ac.exeter.QCRoutines.routines.FixedValue;
 import java.text.ParseException;
 import java.util.List;
 
-import uk.ac.exeter.QCRoutines.Routine;
-import uk.ac.exeter.QCRoutines.RoutineException;
 import uk.ac.exeter.QCRoutines.config.ColumnConfig;
 import uk.ac.exeter.QCRoutines.data.DataRecord;
 import uk.ac.exeter.QCRoutines.data.DataRecordException;
 import uk.ac.exeter.QCRoutines.data.NoSuchColumnException;
 import uk.ac.exeter.QCRoutines.messages.Flag;
+import uk.ac.exeter.QCRoutines.routines.Routine;
+import uk.ac.exeter.QCRoutines.routines.RoutineException;
 import uk.ac.exeter.QCRoutines.util.RoutineUtils;
 
 public class FixedValueRoutine extends Routine {
