+++ src/uk/ac/exeter/QCRoutines/routines/HighDelta/HighDeltaRoutine.java
@@ -5,13 +5,13 @@ import java.util.List;
 import org.joda.time.DateTime;
 import org.joda.time.Seconds;
 
-import uk.ac.exeter.QCRoutines.Routine;
-import uk.ac.exeter.QCRoutines.RoutineException;
 import uk.ac.exeter.QCRoutines.config.ColumnConfig;
 import uk.ac.exeter.QCRoutines.config.ColumnConfigItem;
 import uk.ac.exeter.QCRoutines.data.DataRecord;
 import uk.ac.exeter.QCRoutines.data.DataRecordException;
 import uk.ac.exeter.QCRoutines.data.NoSuchColumnException;
+import uk.ac.exeter.QCRoutines.routines.Routine;
+import uk.ac.exeter.QCRoutines.routines.RoutineException;
 
 public class HighDeltaRoutine extends Routine {
