+++ src/uk/ac/exeter/QCRoutines/routines/Outlier/OutlierRoutine.java
@@ -3,13 +3,13 @@ package uk.ac.exeter.QCRoutines.routines.Outlier;
 import java.util.ArrayList;
 import java.util.List;
 
-import uk.ac.exeter.QCRoutines.Routine;
-import uk.ac.exeter.QCRoutines.RoutineException;
 import uk.ac.exeter.QCRoutines.config.ColumnConfig;
 import uk.ac.exeter.QCRoutines.config.ColumnConfigItem;
 import uk.ac.exeter.QCRoutines.data.DataRecord;
 import uk.ac.exeter.QCRoutines.data.DataRecordException;
 import uk.ac.exeter.QCRoutines.data.NoSuchColumnException;
+import uk.ac.exeter.QCRoutines.routines.Routine;
+import uk.ac.exeter.QCRoutines.routines.RoutineException;
 
 public class OutlierRoutine extends Routine {
