+++ src/uk/ac/exeter/QCRoutines/routines/Routine.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QCRoutines;
+package uk.ac.exeter.QCRoutines.routines;
 
 import java.util.ArrayList;
 import java.util.List;
