+++ src/uk/ac/exeter/QCRoutines/routines/ShipSpeed/ShipSpeedRoutine.java
@@ -4,12 +4,12 @@ import java.util.List;
 
 import org.joda.time.DateTime;
 
-import uk.ac.exeter.QCRoutines.Routine;
-import uk.ac.exeter.QCRoutines.RoutineException;
 import uk.ac.exeter.QCRoutines.config.ColumnConfig;
 import uk.ac.exeter.QCRoutines.data.DataRecord;
 import uk.ac.exeter.QCRoutines.data.DataRecordException;
 import uk.ac.exeter.QCRoutines.messages.Flag;
+import uk.ac.exeter.QCRoutines.routines.Routine;
+import uk.ac.exeter.QCRoutines.routines.RoutineException;
 
 public class ShipSpeedRoutine extends Routine {
